feat: 用户领券和使用模块

+++ coupon-calculation-serv/coupon-calculation-impl/src/main/java/com/yejunyu/coupon/calculation/template/impl/RandomReductionTemplate.java
@@ -19,7 +19,7 @@ public class RandomReductionTemplate extends AbstractRuleTemplate implements Rul
     @Override
     protected Long calculateNewPrice(Long orderTotalAmount, Long shopTotalAmount, Long quota) {
         long maxBenefitAmount = Math.min(quota, shopTotalAmount);
-        long benefitAmount = new Random().nextLong(maxBenefitAmount);
+        long benefitAmount = new Random().nextInt((int) maxBenefitAmount);
         return orderTotalAmount - benefitAmount;
     }
